arm64: configs: Enable ARCH_NXP for standalone config
The issue was that on ls1046 the PCI end node were not found. The root
cause was that the driver for the PCI host side was not enabled. To fix
this issue was to enable ARCH_NXP. As all the dependcies came to this
config option.

While at this, clean the config by running savedefconfig.