net: micrel: Fix the frequency adjustments
The lan841's internal frequency is 125MHz, therefore for each 1 nsec
adjustment of the LTC it contains 8 clock cycles. So the 1PPM format for
lan8841 is wrong, it was using the value of the lan8814 which has
internal frequency of 250.
Therefore fix this by using the correct 1PPM format for lan8841.

Fixes: cafc3662ee3f ("net: micrel: Add PHC support for lan8841")
Signed-off-by: Horatiu Vultur <horatiu.vultur@microchip.com>