arm64: dts: Change the interrupt type for the button used by APPL (DBB-390)
Currently the interrupt type for the push button on both pcbs is level
triggered. This means that if the button is continously pressed or there
is a HW issue, then it would continously generate interrupts and and
after a while it would get out of memory.
The fix consists in changing the interrupt type to be edge trigged, on
both edges, as this is what the HW is expecting.