spi: spidev: Remove matchdata for spidev compatible string
Now, if you declare the compatible string 'spidev' in the DT, that would
fail with the error "spidev listed directly in DT is not supported".
Before this was just a warning.
One solution is to change the compatible string in DT but if there are
any older DT then they will not work anymore.
The other solution, which is implemented here, is to remove this check
for spidev.