dts: lan969x: Fix reboot and reset using push button
Apparently on the board there are small issues regarding the logic for
the reset. In theory we should not be able to toggle GPIO2 to reset the
board. But this works in u-boot.
It doesn't work in the linux kernel because by default the GPIO is set
to output high, so when the GPIO is pulled down or the button is pressed
the voltage on the pin doesn't go to a value of 0.
The fix consists of setting the GPIO as input, then the GPIO doesn't
pull up the line when the button is pressed.
This will be changed in the next revision of the boards