reset: lan969x: Enable RESET_MCHP_SPARX5 for LAN969x
Enable the RESET_MCHP_SPARX5 for lan969x. Currently the driver doesn't
do anything, it is just probed and nothing else. The reason of this is
because when the switch core is reseted the console stops and nothing
seems to be working. This is similar issue with a issue in u-boot which
was fixed by adding a delay after the reset but the same 'fix' doesn't
seem to be working here. Until root-cause the issue don't do anything.
This should not have any sideeffects, as if the linux is booted from
emmc/nor then the network driver in u-boot is not enabled, so it is not
touching the switch core. If the image is downloaded over network in
u-boot then after the image is downloaded it resets the switch.
So in theory the switch should be reseted once it reaches the linux
kernel.