reset: lan969x: Enable the reset of the switch
Before the reset of the switch was not doing anything, it was just an
empty function. The issue was that once the switch was reseted then the
entire chip was stop working. The issue seems to be dissapear once we
change to latest ATF. Therefere we can enable back the switch.