spi: atmel-quadspi: Update for lan969x
Update the driver to support for lan969x. There was some support for
lan966x but the init/setup sequence was different than what was needed
for lan969x. Sometimes this was working on lan969x and sometimes it
didn't work. Therefore update the setup to be specific for lan966x and
this seems to work both on lan969x and lan966x.
Also this allign to what is in u-boot and tfa.