lan969x: Fix the driver for APPL
When was reading the IFH, it was doing a htonl swap, but that looks
wrong. And then also it makes all the offsets to be different than what
is in the documentation.
Therefore stop swapping the bytes for IFH and update the driver to have
the correct offsets for chip port and vlan.