lan969x: Update the network driver for APPL
It was noticed when more traffic was sent to CPU, some of the frames had
the wrong IFH and the API was complaining about this. The issue was that
when a frame was extracted, if already the next frame was in the queue
it would read part of the next frame and add it part of the current
frame. So in this way the IFH of the next frame was getting corrupted.
The fix consists at checking the values that are read and stop when we
reach end of frame. The implementation is taken from the switchdev
driver as that it is working fine.