net: lan969x-switch: add support for FDMA
This commits adds a separate implementation of FDMA for lan969x.

Sparx5 has native coherent memory, therefore migrating to the DMA
mapping API will cause a performance decrease of both RX and TX.

lan969x does *not* have native coherent memory and therefore needs to
use the DMA mapping api to obtain a contigous block of coherent memory.

It was not possible (give the deadline) to come up with a common
solution for both platforms, as one solution would work on lan969x and
not sparx5, and vice versa.