net: lan969x-switch: Fix compilation issue
When kernel was compiled inside the repo, then it failed to compile with
the error:

fatal error: sparx5_main_regs.h: No such file or directory

But it compiled fine when it was compiled outside of the sources.
The reason is that when is compiled outside of the sources, to the
compilation there are added the files to the folder as the include path
but not when is compiled inside the sources.