lan969x: Fix received timestamps when using FDMA
When receiving a frame, the nanosecond part of the timestamp resides in
the IFH. The IFH was parsed and the timestamp was extracted but the
problem was that it was not used. It was always using a local variable
with was not initialized.
Fix this by using the timestamp from the IFH.