dts: lan969x: Add mux for controlling the 1PPS input source
On ev23x71a, it is possible to have multiple sources of 1PPS but only 1
can be active. In can come from the 2 RJ45 or from one SMA connection.
It is possible to control only the 2 RJ45 connections which one to use.
If none of them are enabled then will use SMA connection. It is an issue
to have one RJ45 enabled and also use the SMA connection.
So by default disable the 2 RJ45 and use the SMA connection.