sparx5/lan969x: Update registers
It seems there are also difference regarding the size of fields in the
registers. For example PTP_PIN_SELECT it has a size of 3 on lan969x
while on sparx5 it was a size of 2. This can introduce issues if we
select a pin bigger than 3 on lan969x.
Fix this for all the fields.
We are required to substract 1 from the size because GENMASK includes
both ends otherwise we get bigger masks.