net: sparx5/lan969x: add indirection for target size
The register macros was missing indirection for target size. Target size
is only used to detect if an invalid target is indexed. As an example,
DEV2G5 has size 64 on sparx5 and 28 on lan969x. Before this commits, you
would not get a warning if you tried to index 2g5 device 29 on lan969x.