net: lan969x: get rid of hardcoded calendar
There was a few smaller issues and then the major one that caused
traffic to not work. Basically, the spx5_wr() would override the calendar
choice, in which case the wrong calendar was configured. This happens
because on lan969x, the calendar is chosen using the *PGM_SEL register, but
the choice would get reset by the subsequent spx5_wr() call. Fix this by
using spx5_rmw() instead.

Also assign the correct bandwidth to internal ports. A new ops->
function is added for this purpose.