net: lan969x: Add checks for period and how long it is active
It was noticed that there were missing any checks for how long the
period of a signal it. In the way the user can introduce that the period
is 2 sec and the signal to be high for 1msec. That means that the low
period os 1.9sec but this value can't be written in the registers. As it
can support only 30bit values in nanosecond unit. Therefore add a check
for this and return as invalid these values.