net: micrel: Fix detection of 1PPS for lan8841
When it receives 1PPS input, the lan8841 can be configured to detect the
rising edge or falling edge. The issue is when it detects the edge the
SW will check if any of the edges is detect, but the issue was that the
check was incorrect. Meaning that it was checking for both edges, that
is incorrect, therefore fix this.