serdes: sparx5: Fix powering down the serdes
There are 2 type of lanes as there are 2 types of serdes. One for 25G
and the lowers. The lower ones are from 0 to 24 and this covers both 6G
and 10G and the 25G serdes are from 0 to 7.
The issue is that when powering down the serdes for 6G and 10G it is
using stpidx which is an index inside the type of the serdes and not the
index to the serdes. Meaning that it would power off the first 13
serdes.
The fix consists in using the correct index which is serdes index which
is sidx.