net: lan969x: Add partial support for reference clock
The reference clock is need for PTP configuration because the generated
clock used by PTP timestamping(180 or 328) differs a bit based on the
reference clock. Therefore add reference clock such that we get better
results when adjusting the clock.
Currently all the reference are using a ref clock of 39MHz.