vcap: Fix verifing the keyset
When decoding keyset, it was looking at the mask of the type and if the
mask it was 0, then it return error because it din't know the mask. But
this is not always the case, it can be cases where the type is masked
out to match on multiple keysets. Therefore in this case return OK as it
can match on multiple keysets.