lan966x: vcap: add workaround to hit ipv6 keysets
This commit adds a check for ipv6 SIP and DIP to determine if we should
hit the ipv6 port keysets. Before the merge, port keyset determination
was done differently thatn selecting the rule keyset. With the new VCAP
API both the template keyset and the port keyset is using the same
dissectors, which caused some tests to fail, because adding certain keys
(L3_IP_PROTO) would exclude ipv6 ethtype to match on non-ipv6 keysets.

The check conditionally add the keys required to hit an ipv6 keyset,
when either ipv6 SIP or DIP keys has been added to the list of keys by
the dissectors.

- Add workaround to hit ipv6 keysets
- Add ETH_P_SNAP support
- Add fix for ETH_P_ALL that would only configure the OTHER key
  register. Now configure ALL key registers, when protocol is all