lan966x: vcap: Fix previous workaround for ipv6 keyset hit
This commit gets rid of the circular dependency between basic_usage()
and ipv6_usage() that was introduced in the previous workaround for
hitting the ipv6 keyset.

This solution uses the ipv6_usage() dissector to remove the TCP_IS key
from the rule, if it is present, and instead adds the l4 proto in the
L3_IP_PROTO key, as is supported by the ipv6 keysets. This removes both
the circular dependency and most of the plumbing code to handle it.

 Please enter the commit message for your changes. Lines starting