vcap: Update the reading tc statistics
When using the command 'tc -s filter show dev eth0 ingress' then this
was going to read the rule counters and then set it to 0 because next
time when the tc command it called it needs to return the number of
packets that were hit by this rule from the previous read. All this make
sense. The problem is that when dumping a rule using debugfs then it was
reading the HW counters which were 0 because the tc command clears them.
So then it was showing the wrong information.
The fix consists in not clearing the HW hit count when using tc, but
actually store the last value in SW and then next time when using tc
just subtract from HW counter the previous SW value. In this way it is
not needed any changes to debugfs because the HW hit counter will not be
ever changed.