lan966x: mrp: Fix deletion of MRP ring
When deleting a MRP ring, it was checking the ports but they were not
passed by design to the switchdev object. So they were always null so
the check was always failing. Another problem was that
lan966x_handle_mrp_del was called for each of the ports, meaning that
in case the mrp instance was not found, it was created and then deleted
again.
So the fix consists in returning 0 if the instance is not found and stop
checking the ports.