net: phy: micrel: Extend drivers with fixes and ptp support
net: phy: micrel: Fix force speed 100

It was noticed that if two PCB8290 are connected back to back and force
the ports to run at speed 100, then after few unplug and plug back the
link was not detected.
As this seems to be a HW issue, there is a SW workaround this issue. So
in that case start a background thread and check for the link, if the
link is not up after 5 seconds, then reset the PHY. The reset of the PHY
will reset only the state machines and not the registers meaning that is
not required to configure again the PHY. This seems to fix the issue.

net: phy: micrel: Add support for lan8842

The LAN8842 is a low-power, single port triple-speed (10BASE-T/ 100BASE-TX/
1000BASE-T) ethernet physical layer transceiver (PHY) that supports
transmission and reception of data on standard CAT-5, as well as CAT-5e and
CAT-6, Unshielded Twisted Pair (UTP) cables.

The LAN8842 supports industry-standard SGMII (Serial Gigabit Media
Independent Interface) providing chip-to-chip connection to a Gigabit Ethernet
MAC using a single serialized link (differential pair) in each direction.

There are 2 variants of the lan8842. The one that supports timestamping
(lan8842) and one that doesn't have timestamping (lan8832).

net: phy: micrel: Add PTP support for lan8842

The LAN8842 supports high-accuracy timestamping functions to support
IEEE-1588 solutions using Microchip Ethernet switches, as well as customer
solutions based on SoCs and FPGAs. The LAN8842 also provides Synchronous
Ethernet (SyncE) and TSN frame preemption support.

micrel: lan8814: Fix for skews that don't support 1588

There are skews of Indy that actually don't support 1588. Currently the
driver enables the PHY timestamping for all the skews and that is a
mistake.
Fix this by not allowing PHY timestamping for the skews that don't
support this (8804 and 8808).

micrel: lan8842: Add erratas

Currently there are 2 errats:
1. One for PHY-side shorted center taps.
2. One for Waketx timer which is used for eee.

These values are comming from APPS and they say are similar to the Indy
erratas.

micrel: lan8814: Fix for skews that don't support 1588

In the previous commit that try to fix the skews that don't support 1588
we still allowed to read the PTP time which apparently it was 0. So
instead creating a ptp device which gives wrong information then don't
creat that anymore. Also when the interrupt is handled jump over PTP and
GPIO which are related to PTP.

This is the previous commit:
2b514d629b40 ("micrel: lan8814: Fix for skews that don't support 1588")

net: phy: micrel: Don't create ptp clocks if PHY_TIMESTAPING is not enabled

Already in the function lan8814_ptp_init we get out early if the
PHY_TIMESTAMPING is not enabled, but we still create PTP clock in the
function lan8814_ptp_probe_once. There is no function issue for this
only that is confusing to have those clocks when they are not used.
Therefore fix this by checking if the PHY_TIMESTAMPING is enabled and if
it is not just get out without creating clocks.
Because of this change we need to initialize shared->phydev regardless
if PHY_TIMESTAMPING is enabled or not because when an interrupt will
come we will use this, so if it is not initialized we will get an error.

net: phy: micrel: Fix LED for lan8842

By default the LED will not blink even if there is traffic passing
through the PHY. Therefore change this to show when there is traffic.

phy: micrel: fix register addresses in phy magjack workaround

This workaround is used for 8814, 8842 and 8832. Validation noticed the
address was duplicated instead of incremented in our code.

It is supposed to be incremented, which can be confirmed both from the
PHY datasheets and by looking at the corresponding code in MEPA.

net: phy: add lan9645x internal phy driver

phy: lan8832: add powerdown disable workaround and 8814 workaround

phy: lan9645x: add second internal phy workaround