net: microchip: vcap: fix overflow in count
vcap_get_rule_count_by_cookie must report tc filter pkt counts, since
last request.

To achieve this, without resetting HW counters, this is done:

				ctr->value += temp.value - ri->counter.value;

where the arithmetic is with u32. This would overflow in some
circumstances. In particular, if a vcap rule is disabled, the sw counter
was not reset, leading to overflow.

A consequence of this trick, is that the SW counter cache, ri->counter,
is owned by the tc filter, in the sense that writing to ri->counter.value
when not reporting stats to tc, will break this logic. Therefore, the
updates are removed from the debugfs.