i2c: at91: Set has_clear_cmd to false for sam9x60
It was noticed that even that sam9x60 has the clear command which is
send at the beginning to make sure that the device slave is reseted this
doesn't mean that actually the slave device implements this. Meaning
that the slave device can still pull down the SDA meaning that the bus
will hang if the master is removed during a read.
By setting this to false, master will make sure that will generate
enough clocks (max 9 clocks) such that the slave will finish the
previous transfer. This way of getting the bus out of reset is a common
procedure.