net: sparx5: Enable external detect of signal detect
By default the signal detect is determined by the SGPIO port that is
hardcoded to the chip port. For example if chip port is 24 then SGPIO
port 24 it will say if there is a signal detect or not. But the HW has
the functionality to map chip ports to different SGPIO ports. And this
is done every time when the device tree property microchip,sd-sgpio is
added to the device tree.