net: sparx5: move netdev and notifiers initialization to probe
These two functions registers the net devices and notifier blocks. In
case the sparx5_start() function returns with an error, before the net
devices are registered, they are still attempted untregistered in
probe(). This is a problem.

As it is, we cannot know if the net devices were properly registered, if
the sparx5_start() functions returns an error, because that can happen
in many places. Therefore move the initialization to probe and clean up
in case of any errors.

Signed-off-by: Daniel Machon <daniel.machon@microchip.com>