net: microchip: fix lan966x netdev of_node to enable dsa host usage
Using lan966x as a cpu host for a DSA switch is not possible, because
the netdevices of_node does not correspond to the port of_node in the
DTs.

DSA will find the switch of_node by chasing the dev->parent pointers.
Setting the of_node explicitly fixes this issue.