net: micrel: lan8841: Fix lan8841_handle_interrupt
In the function lan8841_handle_interrupt which is called when the PHY
interrupt line is active, there was a call to
lan8814_handle_gpio_interrupt. This call is wrong because lan8841
handles the GPIO interrupts under PTP interrupt handle.
Then because of this incorrect call, the kernel was crashing because it
was accessing a NULL ptr.