net: phy: micrel: Fix force speed 100
It was noticed that if two PCB8290 are connected back to back and force
the ports to run at speed 100, then after few unplug and plug back the
link was not detected.
As this seems to be a HW issue, there is a SW workaround this issue. So
in that case start a background thread and check for the link, if the
link is not up after 5 seconds, then reset the PHY. The reset of the PHY
will reset only the state machines and not the registers meaning that is
not required to configure again the PHY. This seems to fix the issue.