net: lan969x: add redbox support
The lan969x switch device supports up-to five redboxes, that can be
configured for either: DANH, DANP or HSR-SAN mode. In general, a redbox
forwards frames between its LRE (Link Redundancy Enityty) ports: A, B
and C, depending on the mode.

         lrec
          |
 +-----------------+
 |   Switch core   |
 +-----------------+
          | Interlink (lrec)
 +----------------+
 |     RedBox     |
 +----------------+
     |       |
   lrea    lreb

In hardware, the redbox is implemented as a "mini-switch" with
forwarding masks and lookup tables, to determine the egress port of a
frame. Furthermore, the redbox implements a duplicate discard table, in
order to detect and discard duplicates on LRE ports A or B.

In DANH and DANP modes, frames are not forwarded to the switch core via
the interlink, except if its a supervision frames, in which case it is
redirected to the CPU.

In HSR-SAN mode, frames might be forwarded to the swicth core, via the
interlink, in which case the switch core forwards the frame to the SAN
via normal switch forwarding decisions.

Supervision frames ingressing the LRE ports will be redirected to the
CPU (and might be forwarded in DANH and HSR-SAN modes). The switch does
not send any supervision frames on its own. This is done by the HSR
kernel driver, that periodically sends supervision frames on the LRE
ports.

Some constraints:

- LRE ports A and B (the front ports), must be attached, physically, to
  the same taxi bus in hardware. Configuration will bail out in case
  they are not.

- Redboxes cannot be interconnected.

- Bridged ports and redbox ports cannot overlap. Configuring a redbox
  port, that is already bridged will cause the configuration to bail
  out, and vice-versa.

Signed-off-by: Daniel Machon <daniel.machon@microchip.com>