net: microchip: vcap: add support for u64 sw counter cache
The vcap counters are 32bit, like all other hw counters. To handle
overflow it is necessary to read them sufficiently often, and accumulate
them in a 64bit software cache.

The VCAP api did not support this. The ri->counter on the internal rule
struct was used as a 'last read' value for TC.

However, since the counters are not read systematically on a timer,
overflow was not handled.

This upgrades the struct vcap_counter to be 64bit wide, and renames the
counter used to report 'packets-since-last-check' to TC.

The ->counter field on a rule is now a 64bit accumulating cache of the
hw counts.

Clients can use the vcap_update_counters function on a timer, so make
sure vcap counters are read sufficiently often.