dts: ocelot: Add dyn-cpu-clock
Before the ahb_clock was using a fix clock the cpu_clk but this option
can be changed by the bootloader meaning that it could be a mismatch
between what the u-boot configures and what the kernel expects.
With this driver, it would read the registers of the chip and then set
the correct clock.