net: sparx5: Fix PTP
When we moved the functions around in the probe function, the PTP stop
working because PTP part is initialized before actually seeing if the
PTP threads can be requested. And we allow HW timstamping only after
threads are created, so we have out of order operations in the probe.
The fix consists in requesting PTP threads inside the function
sparx5_ptp_init.