dts: sparx5: Try again to increase I2C bus frequency to 400KHz
The first try to increase the I2C bus frequency to 400KHz was not
successfull as the measure clock frequency was only ~350KHz, even that
it was set in the device tree to be 400KHz.
This time update the i2c-sda/scl-falling-time-ns in this way we can
measure a frequency of 401-402KHz.