hsr: fix hsr_get_node for PRP spv frames or when HW strips htag
The blamed commit introduces a bug in two cases:

1) For PRP supervision frames, which use ETH_P_PRP. They do not contain
a HSR tag, but we still need to maintain the node table based on these
frames. We want to use the sequence number in the RCT for these.

2) When HW strips the tag from an HSR supervision frame, the
SupEtherType (ETH_P_PRP) is what software sees. Again we must handle
these frames instead of bailing out.

Fixes: ddbec99f5857 ("hsr: Fix uninit-value access in hsr_get_node()")