net: sparx5/lan969x: use switchdev_handle_fdb_event_to_device helper
Use the switchdev_handle_fdb_event_to_device helper to fan out FDB
events on bridge- and LAG devices. This is done in Ocelot and lan966x,
and *might* not be necessary on sparx5 and lan969x. I chose to stay
consistent with the other drivers and did it the same way.

This might also come in handy later when introducing support for foreign
devices, which the helper also handles.

DBB-821

Signed-off-by: Daniel Machon <daniel.machon@microchip.com>