hsr: fix: illegal HSR tag modification on forwarding
The IEC 62439-3 ED4 standard of 2021, and all the way back to the 2012
standard, dictate that when forwarding HSR tagged frames between port
A and B, you must not alter the tag.

The rx, tx and forwarding logic hits the same code paths in the driver,
so at the moment the pathid in the htag is always modified to match the
egressing port, which conflicts with the standard in the case of
forwarding.