phy: microchip: lan9645x: add serdes, rgmii and cuphy driver
phy: microchip: lan9645x: Fix set_mode function

When setting the serdes, first it searches if the configuration is
possible by populating the struct needle with different parameters and
then searches through existing possible configurations.
The problem was that after the needle is populated there are checks
which changes the submode, but it was changing the submode of the
argument and not submode of the needle, therefore the needle was never
updated with the new submode so then the checks was failing returning an
error.
Fix this by changing the submode under the needle.

phy: microchip: lan9645x: Fix usage of regmap_update_bits

The function regmap_update_bits has the following declaration:

static inline int regmap_update_bits(struct regmap *map, unsigned int reg,
				     unsigned int mask, unsigned int val);

Where the last argument is the value and the second from last is the
mask but there are 3 usage of this function in this file where the mask and
the val are swapped. Meaning that we don't set necessary the expected value.
Therefore fix this.
We have detected this issue by seeing that the DLL is not getting out of
reset.

phy: lan9645x: add serdes reset and poweroff

This lowers power usage, and with DAC SFPs the link partner will
not register link down unless TX_RESET is asserted.

phy: lan9645x: add serdes calibrate

During link changes the PCS can lose sync to the serdes, which requires
a serdes RX reset to recover from.

phy: lan9645x: add rx/tx reset at start of serdes configuration

Failure to lock requires a RX reset in the serdes. At the moment the
set_mode only applies configuration.

Make sure to perform rx/tx reset at the beginning of the configuration,
which is run as part of the major_mac_config flow.

net: sparx5: Add support for ethtool pause parameters

Implement get_pauseparam() and set_pauseparam() ethtool operations for
Sparx5 ports.  This allows users to query and configure IEEE 802.3x
pause frame settings via:

ethtool -a ethX
ethtool -A ethX rx on|off tx on|off autoneg on|off

The driver delegates pause parameter handling to phylink through
phylink_ethtool_get_pauseparam() and phylink_ethtool_set_pauseparam().

Signed-off-by: Daniel Machon <daniel.machon@microchip.com>