net: sparx5/lan969x: create ISDX pool and use it for PSFP
Currently, ISDX values have only been used for PSFP flows, where the
stream filter id (sfid_id) was mapped 1:1 to ISDX. This will not work
when per-flow BUM policer support is added, as ISDX is also used to
index the BUM policers.

In preparation for BUM policer support, we use the internal pool API to
create a pool of ISDX values, and add a helpers for getting and putting
ISDX values. Update PSFP code in tc_flower, to use the new helpers.

Signed-off-by: Daniel Machon <daniel.machon@microchip.com>