net: phy: mscc: Stop clearing the the UDPv4 checksum for L2 frames
We have noticed that when PHY timestamping is enabled, L2 frames seems to
be modified by changing two 2 bytes with a value of 0. The place were
these 2 bytes seems to be random, meaning that the userspace can ignore
this frame or even worst get the wrong correctien value. This seems to
happen when configuring the HW for IPv4 even that the flow is not
enabled.
These 2 bytes correspond to the UDPv4 checksum and once we don't enable
this for L2 frames then the frame doesn't seem to be changed anymore.