dts: microchip: add dts for lan966x-pcb8385 host for lan9645x
dts: microchip: add oic irq controller to lan966x-pcb8385-lan9645x

dts: microchip: lan9645x update pcb8385 dts for silicon

arm: dts: lan9645x: add i2c, sgpio, i2c-mux for sfps

dsa: lan9645x: add dts for lan9645x evb on lan966x eds2 host

arm: dts: pcb8385 lan9645x fix IRQ

The Ghibli EVB uses GPIO 33 as the external interrupt destination.
This is connected, through the EDS2 connector, to MIIM_INT on the
Maserati EDS2 host board. The MIIN_INT on the board is connected
to GPIO 24 on Maserati, and routed to a PIN (J24) on the board.

To wire the Ghibli OIC up with Maserati we use the Maserati GPIO
controller as irq controller with GPIO 24. The default output
polarity of the OIC is active low.

arm: dts: pcb8385 lan9645x fix read of eeprom from SFP

On Ghibli EVB there are 2 SFPs and it is using the internal clock muxing
to be able to talk with the SFPs. This fix by enabling the i2c
controller and setting the correct alternate mode for the GPIOs.
This fix the SW issue but there is also a HW issue with the EVBs so that
is also needed before this will work.

arm: dts: pcb8385 lan9645x enable SFP ports

On Ghibli EVB there 2 SFP that are connected to Ghibli ports 5 and 6.
Now that it is possible to read the EEPROM of the SFPs enable the ports.

arm: dts: pcb8385 lan9645x add ptp irqs

arm: dts: pcb8385 lan9645x: Enable ALT mode for LEDs GPIO

The lan9645x is controlling the LEDs in the RJ45 connection. The
lan9645x it as a feature where it can blink when there is traffic and it
can turn on when there is link. To enable this feature the lan9645x that
are connected to the LED need to have an alternate mode. Therefore add
this.

arm: dts: pcb8385 lan9645x: Add microchip,led-drive_mode

This this property on the port that needs to have it. As there is a pull
down resistor on the pin.

arm: dts: pcb8385 lan9645x: Remove unused GPIO

The lan9645x_sd0_pins was not used anywhere therefore remove it.

arm: dts: pcb8385 lan9645x add ptp gpios

arm: microchip: fix lan9645x dts to validate against bindings

arm: microchip: lan9645x: Add labels to the ports

These labels will be used later to define the latencies

arm: microchip: lan9645x: Add latencies for the Cu ports

Add latencies for the lan9645x evb board for the Cu ports. Without these
values if a 1m cable is connected back to back and run ptp4l it would
give us delay of ~600ns. With these values we get a value of 0 for 0.5m
cable.
The values were provided by the SQA.

lan9645x dts changes

lan9645x dts changes