configs: mips: Enable CONFIG_PREEMPT
We have noticed that sometimes the kernel takes long time by doing some
work, meaning that so other userspace threads can't be run. And this can
introduce different issues, like for example the PTP, where if the RX
thread it taking more than 375ms to take the frames from the kernel
then PTP module will complain about this because it didn't receive the
frames in the expected time, even if the frames arrive as expected to
the socket queue.
We have noticed that enabling this CONFIG_PREEMPT then we are not seeing
this issue anymore because those threads that can take long time can be
preempted such that a thread with higher priority can run. With this
configuration we have not seen any issues with the PTP module.

APPL-6452