net: phy: mscc: Stop calling netif_rx when having irq locks
According to the description of the function netif_rx, if this is called
from process context mut not disable interrupts but we take a spin lock
and we disable the interrupts. And we noticed that from time to
time then the system was hanging, this depends on the number of frames
received to the CPU.
So, the fix here is not to call netif_rx when interrupts are disabled.
Therefore move the skbs on a different list and call netif_rx when
traversing this list because there the spinlock is not needed. With this
change the system doesn't seem to be hanging anymore.