[sparx5] Use SKB reference counting for PTP tx path
There is a race condition in the TX path of PTP event packets.
When the packet is transmitted, DMA is kicked off and then SKB is freed.
However, in case of PTP packet, SKB is queued for timestamping, and is
freed in the interrupt handler.

If PTP IRQ arrives before the call to sparx5_consume_skb(), it will free
the packet, and then sparx5_consume_skb() will cause Oops, when testing
whether SKB is PTP or not.

This change uses SKB reference counting to ensure SKB is freed regardless
of order of events, and no crash occurs.