arm: dts: pcb8385 lan9645x fix IRQ
The Ghibli EVB uses GPIO 33 as the external interrupt destination.
This is connected, through the EDS2 connector, to MIIM_INT on the
Maserati EDS2 host board. The MIIN_INT on the board is connected
to GPIO 24 on Maserati, and routed to a PIN (J24) on the board.

To wire the Ghibli OIC up with Maserati we use the Maserati GPIO
controller as irq controller with GPIO 24. The default output
polarity of the OIC is active low.