arm: dts: pcb8385 lan9645x fix read of eeprom from SFP
On Ghibli EVB there are 2 SFPs and it is using the internal clock muxing
to be able to talk with the SFPs. This fix by enabling the i2c
controller and setting the correct alternate mode for the GPIOs.
This fix the SW issue but there is also a HW issue with the EVBs so that
is also needed before this will work.