phy: microchip: lan9645x: Fix set_mode function
When setting the serdes, first it searches if the configuration is
possible by populating the struct needle with different parameters and
then searches through existing possible configurations.
The problem was that after the needle is populated there are checks
which changes the submode, but it was changing the submode of the
argument and not submode of the needle, therefore the needle was never
updated with the new submode so then the checks was failing returning an
error.
Fix this by changing the submode under the needle.