dsa: lan9645x: Fix lan9645x_phylink_get_caps
This function should return the capabilities that the port it has. It
has to populate the mac capablities and the interfaces that it supports.
The problem is that for the interfaces it returns what is in the device
tree, which might not be correc, this should return what actually what
HW supports. Therefore fix this by setting the correct interfaces.