arm: dts: pcb8385 lan9645x: Enable ALT mode for LEDs GPIO
The lan9645x is controlling the LEDs in the RJ45 connection. The
lan9645x it as a feature where it can blink when there is traffic and it
can turn on when there is link. To enable this feature the lan9645x that
are connected to the LED need to have an alternate mode. Therefore add
this.