dsa: lan9645x: Introduce device tree property microchip,led-drive_mode
This property is an array which contains first the index of the led that
is connected and then the drive mode. By default the drive mode is open
drain meaning that if there is a pull down resistor then the led will be
turn on no matter what. By changing the drive mode to push pull then it
is possible for the GPIO to pull high so that the led will be turn off
if there is no cable attached.