micrel: lan8814: Fix for skews that don't support 1588
In the previous commit that try to fix the skews that don't support 1588
we still allowed to read the PTP time which apparently it was 0. So
instead creating a ptp device which gives wrong information then don't
creat that anymore. Also when the interrupt is handled jump over PTP and
GPIO which are related to PTP.

This is the previous commit:
2b514d629b40 ("micrel: lan8814: Fix for skews that don't support 1588")