net: sparx5: lan969x: Fix rx timestamping
When the XDP support was introduced the logic for rx timestamping has
been changed. With these change the ifh header was not parsed anumore in
lan969x_fdma_rx_get_frame but in sparx5_fdma_rx_process_frame. And the
problem is that only the src port information was pass and not also the
rx timestamping. So then when the frames were timestamp the rx
timestamp was missing and it always had a value of 0 for the ns part.
The fix consists in passing also the timestamp information from function
sparx5_fdma_rx_process_frame to lan969x_fdma_rx_get_frame.
This issue is not seen on our boards because the PHYs that we have have
timestamp support so the PHY will do the timestamping, to see the issue
it was supposed to disable the PHY timestamping.