spi: spi-vcoreiii: Add delays for clock when doing bit-banging
It was noticed that when the system was under stress by reading rootfs
from NOR and sending and receiving traffic at high speed then sometimes
the decompress of the xz was failing with the following error:

---
[42931.911885] SQUASHFS error: xz decompression failed, data probably corrupt
[42931.919124] SQUASHFS error: Failed to read block 0x4b28b8: -5
[42931.982294] SQUASHFS error: xz decompression failed, data probably corrupt
[42931.989419] SQUASHFS error: Failed to read block 0x4b28b8: -5
[42931.995256] SQUASHFS error: Unable to read data cache entry [4b28b8]
---

The more traffic the faster the issue appeared.

It was noticed also that if the SPI frequency was lowered then the issue
didn't appear. But in that case it was lowered all the way to 1MHz but
then other issues appeared.
So, because we do bit-banging meaning that we change manually the clock,
adding a delay after setting the clock to allow to reach the needed
level, it seemed to be fixing the issue.

APPL-6645