dsa: lan9645x: vcap: remove dynamic enable_dmac
This function is meant to dynamically change the VCAP port configuration
based on the rules written. However, it can lead to broken states, if a
rule is written which configured SMAC, then another rule is written to
the same lookup which configured DMAC.

Now the first rule is broken (will match dmac).