dsa: lan9645x: disable cut through by default
Cut through forwarding has been enabled by default. The kernel
lacks an API to configure cut through, so this was a way to utilize the
feature despite the missing API.

However, cut through forwarding has many drawbacks, which makes it
unsuitable for being enabled by default:

1) DEV[0-8]:MAC_CFG_STATUS:MAC_MAXLEN_CFG can not be respected, because
   frame data is forwarded before it is known how large the frame is.
2) SQA latency tests are last bit out to first bit in. With cut through
   this is negative.
3) CRC errors are forwarded, and must be terminated at destination. For
   same reason as 1).
4) Cut through only works properly for data flows where it is guaranteed
   to be no congestion. Cut through disables the resource management in
   QSYS, which means you get unexpected results during congestion due to
   tail drops.

For now we add a flag to enable it per port. This flag is always false,
and at the moment it is not possible to configure it from user space.